requirements.txt
# 添加异步支持和精确版本
aiohttp==3.9.3
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0